chore: replace `useModalModelMove` with `ModelMoveModal`
