+++ webui/react/src/pages/ModelRegistry.tsx
@@ -13,10 +13,12 @@ import FilterCounter from 'components/FilterCounter';
 import Button from 'components/kit/Button';
 import Empty from 'components/kit/Empty';
 import Input from 'components/kit/Input';
+import { useModal } from 'components/kit/Modal';
 import Tags, { tagsActionHelper } from 'components/kit/Tags';
 import Toggle from 'components/kit/Toggle';
 import Tooltip from 'components/kit/Tooltip';
 import Link from 'components/Link';
+import ModelMoveModal from 'components/ModelMoveModal';
 import Page from 'components/Page';
 import InteractiveTable, {
   ColumnDef,
@@ -36,7 +38,6 @@ import TableFilterSearch from 'components/Table/TableFilterSearch';
 import WorkspaceFilter from 'components/WorkspaceFilter';
 import useModalModelCreate from 'hooks/useModal/Model/useModalModelCreate';
 import useModalModelDelete from 'hooks/useModal/Model/useModalModelDelete';
-import useModalModelMove from 'hooks/useModal/Model/useModalModelMove';
 import usePermissions from 'hooks/usePermissions';
 import { UpdateSettings, useSettings } from 'hooks/useSettings';
 import { paths } from 'routes/utils';
@@ -75,6 +76,7 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
   const loadableUsers = useObservable(usersStore.getUsers());
   const users = Loadable.map(loadableUsers, ({ users }) => users);
   const [models, setModels] = useState<ModelItem[]>([]);
+  const [model, setModel] = useState<ModelItem | null>(null);
   const [tags, setTags] = useState<string[]>([]);
   const [isLoading, setIsLoading] = useState(true);
   const [canceler] = useState(new AbortController());
@@ -89,8 +91,7 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
   const { contextHolder: modalModelDeleteContextHolder, modalOpen: openModelDelete } =
     useModalModelDelete();
 
-  const { contextHolder: modalModelMoveContextHolder, modalOpen: openModelMove } =
-    useModalModelMove();
+  const modelMoveModal = useModal(ModelMoveModal);
 
   const settingConfig = useMemo(() => {
     return settingsConfig(workspace?.id.toString() ?? 'global');
@@ -203,13 +204,6 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
       }
     },
     [fetchModels],
-  );
-
-  const moveModelToWorkspace = useCallback(
-    (model: ModelItem) => {
-      openModelMove(model);
-    },
-    [openModelMove],
   );
 
   const setModelTags = useCallback(
@@ -426,7 +420,8 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
           switchArchived(record);
         },
         [MenuKey.MoveModel]: () => {
-          moveModelToWorkspace(record);
+          setModel(record);
+          modelMoveModal.open();
         },
         [MenuKey.DeleteModel]: () => {
           showConfirmDelete(record);
@@ -453,7 +448,7 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
 
       return { items: menuItems, onClick: onItemClick };
     },
-    [moveModelToWorkspace, showConfirmDelete, switchArchived],
+    [modelMoveModal, showConfirmDelete, switchArchived],
   );
 
   const columns = useMemo(() => {
@@ -612,21 +607,21 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
       },
     ] as ColumnDef<ModelItem>[];
   }, [
-    canDeleteModel,
-    canModifyModel,
+    users,
     nameFilterSearch,
     tableSearchIcon,
     descriptionFilterSearch,
+    workspaceFilterDropdown,
+    workspaces,
     labelFilterDropdown,
     tags,
     userFilterDropdown,
-    users,
+    canModifyModel,
     setModelTags,
+    canDeleteModel,
     ModelActionMenu,
     saveModelDescription,
-    workspaceFilterDropdown,
     workspaceRenderer,
-    workspaces,
   ]);
 
   const handleTableChange = useCallback(
@@ -781,7 +776,7 @@ const ModelRegistry: React.FC<Props> = ({ workspace }: Props) => {
       )}
       {modalModelCreateContextHolder}
       {modalModelDeleteContextHolder}
-      {modalModelMoveContextHolder}
+      {model && <modelMoveModal.Component model={model} />}
     </Page>
   );
 };
